+++ test/examples/example_hpx_channel.cpp
@@ -1,58 +1,8 @@
 #include <hpx/hpx_init.hpp>
-#include <hpx/lcos/local/channel.hpp>
 
-template <typename T, std::size_t PoolSize>
-class Pool {
-  template <class U>
-  class Wrapper {
-    friend class Pool<U, PoolSize>;
+#include "dlaf/common/pool.h"
 
-    Wrapper(hpx::lcos::local::channel<U>* channel, U&& object)
-        : channel_(channel), object_(std::move(object)) {}
-
-  public:
-    Wrapper(Wrapper&& rhs) : channel_(rhs.channel_), object_(std::move(rhs.object_)) {
-      rhs.channel_ = nullptr;
-    }
-
-    ~Wrapper() {
-      if (channel_)
-        channel_->set(std::move(object_));
-    }
-
-    U& get_value() {
-      return object_;
-    }
-
-  private:
-    U object_;
-    hpx::lcos::local::channel<U>* channel_;
-  };
-
-public:
-  Pool() {
-    for (int i = 0; i < PoolSize; ++i)
-      channel_.set(T{});
-  }
-
-  ~Pool() {
-    channel_.close(/*true*/);  // TODO check what force_delete does
-    for (int i = 0; i < PoolSize; ++i)
-      channel_.get().get();
-  }
-
-  hpx::future<Wrapper<T>> get() {
-    return channel_.get().then(hpx::launch::sync,
-        hpx::util::unwrapping(std::bind(&Pool::make_wrapper, this, std::placeholders::_1)));
-  }
-
-private:
-  Wrapper<T> make_wrapper(T&& object) {
-    return Wrapper<T>{&channel_, std::move(object)};
-  }
-
-  hpx::lcos::local::channel<T> channel_;
-};
+using dlaf::common::Pool;
 
 int hpx_main(int argc, char* argv[]) {
   Pool<int, 2> pool;
